+++ lib/moco/entities/project.rb
@@ -15,8 +15,21 @@ module MOCO
 
     # Fetches tasks associated with this project.
     def tasks
-      # Use the has_many method to fetch tasks
-      has_many(:tasks)
+      # Check if tasks are already loaded in attributes
+      if attributes[:tasks].is_a?(Array) && attributes[:tasks].all? { |t| t.is_a?(MOCO::Task) }
+        # If tasks are already loaded, create a NestedCollectionProxy with the loaded tasks
+        @_tasks_proxy ||= begin
+          require_relative "../nested_collection_proxy"
+          proxy = MOCO::NestedCollectionProxy.new(client, self, :tasks, "Task")
+          # We need to manually set the loaded records since we already have them
+          proxy.instance_variable_set(:@records, attributes[:tasks])
+          proxy.instance_variable_set(:@loaded, true)
+          proxy
+        end
+      else
+        # Otherwise, use has_many with nested=true
+        has_many(:tasks, nil, nil, true)
+      end
     end
 
     def active?